- 64 bit updates (4): compiler and debugger plugin
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@7828 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/debuggergdb/debugger_defs.h
@@ -221,8 +221,10 @@ struct StackFrame
         line.Clear();
     }
     bool valid; ///< Is this stack frame valid?
-    unsigned long int number; ///< Stack frame's number (used in backtraces).
-    unsigned long int address; ///< Stack frame's address.
+    // 64 bit compatibility: don't use unsigned long int here:
+    size_t number; ///< Stack frame's number (used in backtraces).
+    // ...and here:
+    size_t address; ///< Stack frame's address.
     wxString function; ///< Current function name.
     wxString file; ///< Current file.
     wxString line; ///< Current line in file.
